[AO] test 404 error page
