--- test/uk/gov/hmrc/homeofficesettledstatus/journey/InMemoryStore.scala
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.homeofficesettledstatus.journey
-
-import java.util.concurrent.atomic.AtomicReference
-
-import scala.concurrent.{ExecutionContext, Future}
-
-/**
-  * Basic in-memory store used to test journeys.
-  */
-trait InMemoryStore[S, C] {
-
-  private val state = new AtomicReference[Option[S]](None)
-
-  def fetch(implicit requestContext: C, ec: ExecutionContext): Future[Option[S]] =
-    Future.successful(state.get)
-
-  def save(newState: S)(implicit requestContext: C, ec: ExecutionContext): Future[S] = Future {
-    state.set(Some(newState))
-    newState
-  }
-}
